change the setup fixture scope to session , seeting up class level driver instance to driver object return from setup fixture
